+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Routing V8: add new transport mode "bicycle" and "scooter";
 - Routing V8: add alternatives parameter;
 - Routing V8: add units (metrics/imperial);
+- Routing V8: add via parameter;
 - Adding test suite
 
 ### Change